+++ optlite-components/js/opt-frontend.ts
@@ -497,7 +497,7 @@ export class OptFrontend extends AbstractBaseFrontend {
   }
 
   handleUncaughtException(trace) {
-    if (trace.length == 1 && trace[0].line) {
+    if (trace.length == 1 && trace[0].line && !trace[0].advisory) {
       var errorLineNo = trace[0].line - 1; /* Ace lines are zero-indexed */
       if (errorLineNo !== undefined 
         // && errorLineNo != NaN